api/requirements.txt
fastapi
uvicorn[standard]
websockets>=12.0
strawberry-graphql[fastapi]>=0.220.0
arq>=0.25.0
requests
sqlalchemy[asyncio]>=2.0.0
pydantic
pydantic-settings
redis>=4.5.0
psycopg2-binary
asyncpg>=0.29.0
greenlet>=3.0.0
alembic
mako
psutil
uvloop
httptools
prometheus_client>=0.17.0
python-jose[cryptography]
python-multipart
passlib[bcrypt]
slowapi
fastapi-limiter
httpx
